# api/swagger.yaml
openapi: 3.1.0
info:
  title: SOSO 配車サポート API
  version: 0.1.0
  description: |
    SOSO配車サービスAPI。  
    認証は「アクセスJWT(Authorizationヘッダ)」＋「リフレッシュトークン(HttpOnly Cookie: rt)」の二段構成。  
    変更系エンドポイントでは CSRF トークン（X-CSRF-Token ヘッダ）を要求します。
servers:
  - url: http://localhost:8080
    description: ローカル開発用サーバ
  - url: http://localhost:8080/api
    description: /api プレフィックスを付ける構成の場合（必要に応じてパスも合わせる）
...
paths:
  /csrf:
    get:
      tags: [System]
      summary: CSRFトークン取得
      description: CSRFトークンを返却し、必要に応じてCookieにも設定します。
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/login:
    post:
      tags: [Auth]
      summary: ユーザーログイン
      description: |
        ユーザーを認証し、アクセスJWTを返却。リフレッシュトークンは HttpOnly Cookie `rt` に設定されます。
      security:
        - CsrfToken: []    # CSRFミドルウェア有効時
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 認証成功
          headers:
            Set-Cookie:
              description: HttpOnly refresh token cookie (rt)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: アクセストークン再発行
      description: |
        有効なリフレッシュトークンCookieを用いてアクセストークンを再発行し、リフレッシュトークンもローテーションします。
      security:
        - RefreshCookie: []
        - CsrfToken: []
      responses:
        '200':
          description: 再発行成功
          headers:
            Set-Cookie:
              description: Rotated HttpOnly refresh token cookie (rt)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: ログアウト
      description: リフレッシュトークンを無効化し、Cookieを削除します。
      security:
        - RefreshCookie: []
        - CsrfToken: []
      responses:
        '204':
          description: ログアウト成功（No Content）
          headers:
            Set-Cookie:
              description: クリアされたリフレッシュトークンCookie
              schema:
                type: string
        '500':
          $ref: '#/components/responses/ServerError'
  /users/register:
    post:
      tags: [Users]
      summary: ユーザー登録
      description: 新規ユーザー登録
      operationId: createUser
      security:
        - CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/me:
    get:
      tags: [Users]
      summary: 自分のユーザー情報取得
      security:
        - AccessToken: []
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # 以降は既存設計（未実装含む）
  /users:
    get:
      tags: [Users]
      summary: ユーザー一覧取得
      description: ユーザー一覧取得（取得数制限付き）
      operationId: listUsers
      security:
        - AccessToken: []
      parameters:
        - name: limit
          in: query
          description: 取得上限件数
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: 取得開始オフセット
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCount:
                    type: integer
                  count:
                    type: integer
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookie:
      type: apiKey
      in: cookie
      name: rt
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
    BadRequest:
      description: 入力値不正エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: 認証情報不正エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: 権限不足エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 競合エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 対象が存在しない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: サーバ内部エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    CsrfResponse:
      type: object
      properties:
        csrf_token:
          type: string
      required: [csrf_token]
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: ログインID
          example: kitano_masaki
        password:
          type: string
          format: password
          description: ログイン用パスワード
          example: Passw0rd!123
    TokenResponse:
      type: object
      required: [access_token, access_expires_at]
      properties:
        access_token:
          type: string
          description: アクセスJWT
        access_expires_at:
          type: string
          format: date-time
          description: アクセスJWTの有効期限
    UserCreateRequest:
      type: object
      required: [username, password, hasCar]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          example: kitano_masaki
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: Passw0rd!123
        hasCar:
          type: boolean
          example: true
        capacity:
          type: integer
          format: int32
          minimum: 0
          example: 4
...
    UserResponse:
      type: object
      required: [id, username, hasCar]
      properties:
        id:
          type: string
          example: usr_12345
        username:
          type: string
          example: kitano_masaki
        hasCar:
          type: boolean
          example: true
        capacity:
          type: integer
          format: int32
          example: 4
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
        sosoPoints:
          type: integer
          format: int32
          example: 120
...